+++ admin/update-service.php
@@ -1,10 +1,8 @@
-<?php include("partials/menu.php") ?> ;
+<?php include("partials/menu.php"); ?> 
 
 <section id="content">
 
     <h2>Update Service</h2>
-    <br>
-
             <?php
 
                     if(isset($_SESSION['upload']))
@@ -17,11 +15,21 @@
                         echo $_SESSION['remove'] ;
                         unset($_SESSION['remove']) ;
                     }          
+                    if(isset($_SESSION['too_large']))
+                    {
+                        echo $_SESSION['too_large'] ;
+                        unset($_SESSION['too_large']) ;
+                    }          
+                    if(isset($_SESSION['format']))
+                    {
+                        echo $_SESSION['format'] ;
+                        unset($_SESSION['format']) ;
+                    }          
             ?>
     <br>    
 
     <div class="container">
-
+        <a href="<?php echo SITEURL; ?>admin/manage-services.php"><button class="btn btn-outline-dark">Back</button></a><br><br>
     <?php
 
         if(isset($_GET['id']))
@@ -64,25 +72,24 @@
     ?>
 
         <form action="" method="POST" enctype="multipart/form-data">
-
-            <table class = "tbl_30">
-
+        <div class="table-responsive">
+            <table class="table table-borderless" style="width:450px;">
                 <tr>
                     <td>Title:</td>
                     <td>
-                        <input type="text" name="title" value = <?php echo $title; ?>>
+                        <input type="text" name="title" class="form-control" value = <?php echo $title; ?> required>
                     </td>
                 </tr>
                 <tr>
                     <td>Description:</td>
                     <td>
-                        <textarea name="description" cols="30" rows="10"><?php echo $description; ?> </textarea>
+                        <textarea name="description" class="form-control" cols="30" rows="10" required><?php echo $description; ?> </textarea>
                     </td>
                 </tr>
                 <tr>
                     <td>Price:</td>
                     <td>
-                        <input type="number" name="price" value = <?php echo $price; ?>>
+                        <input type="number" name="price" class="form-control" value = <?php echo $price; ?> required>
                     </td>
                 </tr>
 
@@ -106,13 +113,13 @@
                 <tr>
                     <td>Upload new image:</td>
                     <td>
-                        <input type="file" name="image">
+                        <input type="file" name="image" class="form-control">
                     </td>
                 </tr>
                 <tr>
                     <td>Category:</td>
                     <td>
-                           <select name="category_id" >
+                           <select name="category_id" class="form-select">
 
                                     <?php
                                             $sql2 = "SELECT * FROM tbl_s_category WHERE active='Yes'" ;
@@ -135,32 +142,33 @@
                 <tr>
                     <td>Featured:</td>
                     <td>
-                        <input type="radio" name="featured" <?php if($featured=='Yes') {echo"checked";} ?> value="Yes">Yes
-                        <input type="radio" name="featured" <?php if($featured=='No') {echo"checked";} ?> value="No">No
+                        <input type="radio" name="featured" <?php if($featured=='Yes') {echo"checked";} ?> value="Yes"> Yes
+                        <input type="radio" name="featured" <?php if($featured=='No') {echo"checked";} ?> value="No"> No
                     </td>
                 </tr>
                 <tr>
                     <td>Active:</td>
                     <td>
-                        <input type="radio" name="active" <?php if($active=='Yes') {echo"checked";} ?> value="Yes">Yes
-                        <input type="radio" name="active" <?php if($active=='No') {echo"checked";} ?> value="No">No
+                        <input type="radio" name="active" <?php if($active=='Yes') {echo"checked";} ?> value="Yes"> Yes
+                        <input type="radio" name="active" <?php if($active=='No') {echo"checked";} ?> value="No"> No
                     </td>
                 </tr>
                 <tr>                    
                     <td colspan="2">
                         <input type="hidden" name="id" value="<?php echo $product_id; ?>">
                         <input type="hidden" name="" value="<?php echo $current_image; ?>">
-                        <input type="submit" name="submit" value="Update Product" class="btn btn-secondary">
+                        <input type="submit" name="submit" value="DONE" class="btn btn-secondary">
                     </td>
                 </tr>
             </table>
+        </div>
         </form>
 
         <?php
                 if(isset($_POST['submit']))
                 {
-                    $title = $_POST['title'] ;
-                    $description = $_POST['description'] ;
+                    $title = mysqli_escape_string($conn, $_POST['title']) ;
+                    $description = mysqli_escape_string($conn, $_POST['description']) ;
                     $price = $_POST['price'] ;
                     $featured = $_POST['featured'] ;
                     $active = $_POST['active'] ;
@@ -168,56 +176,65 @@
 
                     if(isset($_FILES['image']['name']))
                     {
-                        $image_name = $_FILES['image']['name'] ;
-
-                        // Upload
-                        
-                        if($image_name != "")
+                        $image_name = $_FILES['image']['name'] ;                        
+                        if($_FILES['image']['size'] > 1000000)
                         {
-                            $ext = end(explode('.',$image_name)) ;
-                        
-                            $date = time() ;
-
-                            $image_name = "Service-".$date.".".$ext ;
-
-                            $s_path = $_FILES['image']['tmp_name'] ;
-
-                            $d_path = "../img/services/".$image_name ;
-
-                                $upload = move_uploaded_file($s_path, $d_path) ;
-
-                                if($upload==FALSE)
-                                {
-                                    $_SESSION['upload'] = "<div class='error'>!! Failed to Upload Image !!</div>" ;
-                                    header("location:".SITEURL."admin/update-service.php?id=$service_id") ;
-                                    die() ;
-                                }
-                                
-                                    if($current_image != "")
-                                    {
-                                        $r_path = "../img/products/".$current_image ;
-
-                                        $remove = unlink($r_path) ;
-
-                                        if($remove==FALSE)
-                                        {
-                                            $_SESSION['remove'] = "<div class='error'>!! Failed to Remove Image !!</div>" ;
-                                            header("location:".SITEURL."admin/update-product.php?id=$service_id") ;
-                                            die() ;
-                                        }
-                                    }
+                            $_SESSION['too_large'] = "<div class='error'>!! Sorry, your file is too large !!</div>" ;
+                            header("location:".SITEURL."admin/update-service.php?id=$service_id") ;
+                            die() ;
                         }
                         else
                         {
-                            $image_name = $current_image ;
-                        }
+                            // Upload                        
+                            if($image_name != "")
+                            {
+                                $ext = end(explode('.',$image_name)) ;
+                                $allowed_ext = array("jpg", "png", "jpeg") ;
+                                if(in_array($ext, $allowed_ext))
+                                {
+                                    $date = time() ;
 
-                        
-
-                        
-
-                        //echo "biren" ;
-                        //die() ;
+                                    $image_name = "Service-".$date.".".$ext ;
+    
+                                    $s_path = $_FILES['image']['tmp_name'] ;
+    
+                                    $d_path = "../img/services/".$image_name ;
+    
+                                        $upload = move_uploaded_file($s_path, $d_path) ;
+    
+                                        if($upload==FALSE)
+                                        {
+                                            $_SESSION['upload'] = "<div class='error'>!! Failed to Upload Image !!</div>" ;
+                                            header("location:".SITEURL."admin/update-service.php?id=$service_id") ;
+                                            die() ;
+                                        }
+                                        
+                                            if($current_image != "")
+                                            {
+                                                $r_path = "../img/products/".$current_image ;
+    
+                                                $remove = unlink($r_path) ;
+    
+                                                if($remove==FALSE)
+                                                {
+                                                    $_SESSION['remove'] = "<div class='error'>!! Failed to Remove Image !!</div>" ;
+                                                    header("location:".SITEURL."admin/update-product.php?id=$service_id") ;
+                                                    die() ;
+                                                }
+                                            }
+                                }
+                                else
+                                {
+                                    $_SESSION['format'] = "<div class='error'>!! You can't upload files of this type !!</div>" ;
+                                    header("location:".SITEURL."admin/update-service.php?id=$service_id") ;
+                                    die() ;
+                                }                                
+                            }
+                            else
+                            {
+                                $image_name = $current_image ;
+                            }
+                        }                                       
                     }
                     else
                     {
@@ -259,4 +276,4 @@
     </div>
 </section>
 
-<?php include("partials/footer.php") ?> ;
+<?php include("partials/footer.php"); ?>
